fix: adjust api for actual amount

+++ api_server/src/validation/transactionsValidation.ts
@@ -1,14 +1,14 @@
-import { z } from 'zod';
+import { z } from "zod";
 
 const SubtransactionSchema = z.object({
-  amount: z.number().int(),
+  amount: z.number(),
   category: z.string().optional(),
   notes: z.string().optional(),
 });
 
 const TransactionSchema = z.object({
-  date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, 'Must be in YYYY-MM-DD format'),
-  amount: z.number().int().optional(),
+  date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, "Must be in YYYY-MM-DD format"),
+  amount: z.number().optional(),
   payee: z.string().nullable().optional(),
   payee_name: z.string().optional(),
   imported_payee: z.string().optional(),
@@ -27,15 +27,17 @@ const ImportOptsSchema = z.object({
 });
 
 export const AccountIdParamsSchema = z.object({
-  accountId: z.string().min(1, 'accountId is required'),
+  accountId: z.string().min(1, "accountId is required"),
 });
 
 export const ImportTransactionsBodySchema = z.object({
   transactions: z
     .array(TransactionSchema)
-    .min(1, 'At least one transaction is required'),
+    .min(1, "At least one transaction is required"),
   opts: ImportOptsSchema.optional(),
 });
 
 export type AccountIdParams = z.infer<typeof AccountIdParamsSchema>;
-export type ImportTransactionsBody = z.infer<typeof ImportTransactionsBodySchema>;
+export type ImportTransactionsBody = z.infer<
+  typeof ImportTransactionsBodySchema
+>;
